Refactor storage class validation to better match the future webhook
