+++ pkg/apis/elasticsearch/v1/name.go
@@ -57,8 +57,8 @@ var (
 	}
 )
 
-// validateNames checks the validity of resource names that will be generated by the given Elasticsearch object.
-func validateNames(es *Elasticsearch) error {
+// ValidateNames checks the validity of resource names that will be generated by the given Elasticsearch object.
+func ValidateNames(es Elasticsearch) error {
 	if len(es.Name) > common_name.MaxResourceNameLength {
 		return errors.Errorf("name exceeds maximum allowed length of %d", common_name.MaxResourceNameLength)
 	}
