+++ pkg/controller/elasticsearch/driver/pvc_expansion.go
@@ -6,20 +6,16 @@ package driver
 
 import (
 	"encoding/json"
-	"errors"
-	"fmt"
 	"strings"
 
 	esv1 "github.com/elastic/cloud-on-k8s/pkg/apis/elasticsearch/v1"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/elasticsearch/sset"
+	"github.com/elastic/cloud-on-k8s/pkg/controller/elasticsearch/validation"
 	"github.com/elastic/cloud-on-k8s/pkg/utils/k8s"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	storagev1 "k8s.io/api/storage/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes/scheme"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
@@ -45,19 +41,27 @@ func handleVolumeExpansion(
 	actualSset appsv1.StatefulSet,
 	validateStorageClass bool,
 ) (bool, error) {
-	err := resizePVCs(k8sClient, es, expectedSset, actualSset, validateStorageClass)
+	// ensure there are no incompatible storage size modification
+	if err := validation.ValidateClaimsStorageUpdate(
+		k8sClient,
+		actualSset.Spec.VolumeClaimTemplates,
+		expectedSset.Spec.VolumeClaimTemplates,
+		validateStorageClass); err != nil {
+		return false, err
+	}
+
+	// resize all PVCs that can be resized
+	err := resizePVCs(k8sClient, es, expectedSset, actualSset)
 	if err != nil {
 		return false, err
 	}
 
-	recreate, err := needsRecreate(k8sClient, expectedSset, actualSset, validateStorageClass)
-	if err != nil {
-		return false, err
+	// schedule the StatefulSet for recreation if needed
+	if needsRecreate(expectedSset, actualSset) {
+		return true, annotateForRecreation(k8sClient, es, actualSset, expectedSset.Spec.VolumeClaimTemplates)
 	}
-	if !recreate {
-		return false, nil
-	}
-	return true, annotateForRecreation(k8sClient, es, actualSset, expectedSset.Spec.VolumeClaimTemplates)
+
+	return false, nil
 }
 
 // resizePVCs updates the spec of all existing PVCs whose storage requests can be expanded,
@@ -68,7 +72,6 @@ func resizePVCs(
 	es esv1.Elasticsearch,
 	expectedSset appsv1.StatefulSet,
 	actualSset appsv1.StatefulSet,
-	validateStorageClass bool,
 ) error {
 	// match each existing PVC with an expected claim, and decide whether the PVC should be resized
 	actualPVCs, err := sset.RetrieveActualPVCs(k8sClient, actualSset)
@@ -81,27 +84,19 @@ func resizePVCs(
 			continue
 		}
 		for _, pvc := range pvcs {
-			pvcSize := pvc.Spec.Resources.Requests.Storage()
-			claimSize := expectedClaim.Spec.Resources.Requests.Storage()
-			// is it a storage increase?
-			isExpansion, err := isStorageExpansion(claimSize, pvcSize)
-			if err != nil {
-				return err
-			}
-			if !isExpansion {
+			storageCmp := k8s.CompareStorageRequests(pvc.Spec.Resources, expectedClaim.Spec.Resources)
+			if !storageCmp.Increase {
+				// not an increase, nothing to do
 				continue
 			}
-			// is volume expansion supported?
-			if err := ensureClaimSupportsExpansion(k8sClient, *expectedClaim, validateStorageClass); err != nil {
-				return err
-			}
 
+			newSize := expectedClaim.Spec.Resources.Requests.Storage()
 			log.Info("Resizing PVC storage requests. Depending on the volume provisioner, "+
 				"Pods may need to be manually deleted for the filesystem to be resized.",
-				"namespace", pvc.Namespace, "es_name", es.Name,
-				"pvc_name", pvc.Name,
-				"old_value", pvcSize.String(), "new_value", claimSize.String())
-			pvc.Spec.Resources.Requests[corev1.ResourceStorage] = *claimSize
+				"namespace", pvc.Namespace, "es_name", es.Name, "pvc_name", pvc.Name,
+				"old_value", pvc.Spec.Resources.Requests.Storage().String(), "new_value", newSize.String())
+
+			pvc.Spec.Resources.Requests[corev1.ResourceStorage] = *newSize
 			if err := k8sClient.Update(&pvc); err != nil {
 				return err
 			}
@@ -135,54 +130,18 @@ func annotateForRecreation(
 }
 
 // needsRecreate returns true if the StatefulSet needs to be re-created to account for volume expansion.
-// An error is returned if volume expansion is required but claims are incompatible.
-func needsRecreate(
-	k8sClient k8s.Client,
-	expectedSset appsv1.StatefulSet,
-	actualSset appsv1.StatefulSet,
-	validateStorageClass bool,
-) (bool, error) {
-	recreate := false
-	// match each expected claim with an actual existing one: we want to return true
-	// if at least one claim has increased storage reqs
-	// however we want to error-out if any claim has an incompatible storage req
+func needsRecreate(expectedSset appsv1.StatefulSet, actualSset appsv1.StatefulSet) bool {
 	for _, expectedClaim := range expectedSset.Spec.VolumeClaimTemplates {
 		actualClaim := sset.GetClaim(actualSset.Spec.VolumeClaimTemplates, expectedClaim.Name)
 		if actualClaim == nil {
 			continue
 		}
-		isExpansion, err := isStorageExpansion(expectedClaim.Spec.Resources.Requests.Storage(), actualClaim.Spec.Resources.Requests.Storage())
-		if err != nil {
-			return false, err
+		storageCmp := k8s.CompareStorageRequests(actualClaim.Spec.Resources, expectedClaim.Spec.Resources)
+		if storageCmp.Increase {
+			return true
 		}
-		if !isExpansion {
-			continue
-		}
-		if err := ensureClaimSupportsExpansion(k8sClient, expectedClaim, validateStorageClass); err != nil {
-			return false, err
-		}
-		recreate = true
 	}
-
-	return recreate, nil
-}
-
-// isStorageExpansion returns true if actual is higher than expected.
-// Decreasing storage size is unsupported: an error is returned if expected < actual.
-func isStorageExpansion(expectedSize *resource.Quantity, actualSize *resource.Quantity) (bool, error) {
-	if expectedSize == nil || actualSize == nil {
-		// not much to compare if storage size is unspecified
-		return false, nil
-	}
-	switch expectedSize.Cmp(*actualSize) {
-	case 0: // same size
-		return false, nil
-	case -1: // decrease
-		return false, fmt.Errorf("decreasing storage size is not supported, "+
-			"but an attempt was made to resize from %s to %s", actualSize.String(), expectedSize.String())
-	default: // increase
-		return true, nil
-	}
+	return false
 }
 
 // recreateStatefulSets re-creates StatefulSets as specified in Elasticsearch annotations, to account for
@@ -334,66 +293,4 @@ func updatePods(k8sClient k8s.Client, statefulSet appsv1.StatefulSet, updateFunc
 		}
 	}
 	return nil
-}
-
-// ensureClaimSupportsExpansion inspects whether the storage class referenced by the claim
-// allows volume expansion, and returns an error if it doesn't.
-func ensureClaimSupportsExpansion(k8sClient k8s.Client, claim corev1.PersistentVolumeClaim, validateStorageClass bool) error {
-	if !validateStorageClass {
-		log.V(1).Info("Skipping storage class validation")
-		return nil
-	}
-	sc, err := getStorageClass(k8sClient, claim)
-	if err != nil {
-		return err
-	}
-	if !allowsVolumeExpansion(sc) {
-		return fmt.Errorf("claim %s (storage class %s) does not support volume expansion", claim.Name, sc.Name)
-	}
-	return nil
-}
-
-// getStorageClass returns the storage class specified by the given claim,
-// or the default storage class if the claim does not specify any.
-func getStorageClass(k8sClient k8s.Client, claim corev1.PersistentVolumeClaim) (storagev1.StorageClass, error) {
-	if claim.Spec.StorageClassName == nil || *claim.Spec.StorageClassName == "" {
-		return getDefaultStorageClass(k8sClient)
-	}
-	var sc storagev1.StorageClass
-	if err := k8sClient.Get(types.NamespacedName{Name: *claim.Spec.StorageClassName}, &sc); err != nil {
-		return storagev1.StorageClass{}, fmt.Errorf("cannot retrieve storage class: %w", err)
-	}
-	return sc, nil
-}
-
-// getDefaultStorageClass returns the default storage class in the current k8s cluster,
-// or an error if there is none.
-func getDefaultStorageClass(k8sClient k8s.Client) (storagev1.StorageClass, error) {
-	var scs storagev1.StorageClassList
-	if err := k8sClient.List(&scs); err != nil {
-		return storagev1.StorageClass{}, err
-	}
-	for _, sc := range scs.Items {
-		if isDefaultStorageClass(sc) {
-			return sc, nil
-		}
-	}
-	return storagev1.StorageClass{}, errors.New("no default storage class found")
-}
-
-// isDefaultStorageClass inspects the given storage class and returns true if it is annotated as the default one.
-func isDefaultStorageClass(sc storagev1.StorageClass) bool {
-	if len(sc.Annotations) == 0 {
-		return false
-	}
-	if sc.Annotations["storageclass.kubernetes.io/is-default-class"] == "true" ||
-		sc.Annotations["storageclass.beta.kubernetes.io/is-default-class"] == "true" {
-		return true
-	}
-	return false
-}
-
-// allowsVolumeExpansion returns true if the given storage class allows volume expansion.
-func allowsVolumeExpansion(sc storagev1.StorageClass) bool {
-	return sc.AllowVolumeExpansion != nil && *sc.AllowVolumeExpansion
 }
