+++ pkg/controller/elasticsearch/driver/pvc_expansion_test.go
@@ -7,7 +7,6 @@ package driver
 import (
 	"encoding/json"
 	"fmt"
-	"reflect"
 	"testing"
 
 	esv1 "github.com/elastic/cloud-on-k8s/pkg/apis/elasticsearch/v1"
@@ -45,14 +44,6 @@ var (
 			Resources: corev1.ResourceRequirements{Requests: map[corev1.ResourceName]resource.Quantity{
 				corev1.ResourceStorage: resource.MustParse("1Gi"),
 			}}}}
-
-	defaultStorageClass = storagev1.StorageClass{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        "default-sc",
-			Annotations: map[string]string{"storageclass.kubernetes.io/is-default-class": "true"}}}
-	defaultBetaStorageClass = storagev1.StorageClass{ObjectMeta: metav1.ObjectMeta{
-		Name:        "default-beta-sc",
-		Annotations: map[string]string{"storageclass.beta.kubernetes.io/is-default-class": "true"}}}
 
 	sampleSset = appsv1.StatefulSet{ObjectMeta: metav1.ObjectMeta{Namespace: "ns", Name: "sample-sset"}}
 )
@@ -240,165 +231,44 @@ func Test_handleVolumeExpansion(t *testing.T) {
 
 func Test_needsRecreate(t *testing.T) {
 	type args struct {
-		k8sClient            k8s.Client
-		expectedSset         appsv1.StatefulSet
-		actualSset           appsv1.StatefulSet
-		validateStorageClass bool
+		expectedSset appsv1.StatefulSet
+		actualSset   appsv1.StatefulSet
 	}
 	tests := []struct {
-		name    string
-		args    args
-		want    bool
-		wantErr bool
+		name string
+		args args
+		want bool
 	}{
 		{
 			name: "requested storage increase in the 2nd claim: recreate",
 			args: args{
-				k8sClient:            k8s.WrappedFakeClient(withVolumeExpansion(sampleStorageClass)),
-				expectedSset:         withClaims(sampleSset, sampleClaim, withStorageReq(sampleClaim2, "3Gi")),
-				actualSset:           withClaims(sampleSset, sampleClaim, sampleClaim2),
-				validateStorageClass: true,
+				expectedSset: withClaims(sampleSset, sampleClaim, withStorageReq(sampleClaim2, "3Gi")),
+				actualSset:   withClaims(sampleSset, sampleClaim, sampleClaim2),
 			},
 			want: true,
 		},
 		{
-			name: "volume expansion not supported: error out",
-			args: args{
-				k8sClient:            k8s.WrappedFakeClient(&sampleStorageClass),
-				expectedSset:         withClaims(sampleSset, sampleClaim, withStorageReq(sampleClaim2, "3Gi")),
-				actualSset:           withClaims(sampleSset, sampleClaim, sampleClaim2),
-				validateStorageClass: true,
-			},
-			want:    false,
-			wantErr: true,
-		},
-		{
-			name: "volume expansion not supported but no storage class validation: recreate",
-			args: args{
-				k8sClient:            k8s.WrappedFakeClient(&sampleStorageClass),
-				expectedSset:         withClaims(sampleSset, sampleClaim, withStorageReq(sampleClaim2, "3Gi")),
-				actualSset:           withClaims(sampleSset, sampleClaim, sampleClaim2),
-				validateStorageClass: false,
-			},
-			want:    true,
-			wantErr: false,
-		},
-		{
 			name: "no claim in the StatefulSet",
 			args: args{
-				k8sClient:            k8s.WrappedFakeClient(withVolumeExpansion(sampleStorageClass)),
-				expectedSset:         sampleSset,
-				actualSset:           sampleSset,
-				validateStorageClass: true,
+				expectedSset: sampleSset,
+				actualSset:   sampleSset,
 			},
 			want: false,
 		},
 		{
 			name: "no change in the claim",
 			args: args{
-				k8sClient:            k8s.WrappedFakeClient(withVolumeExpansion(sampleStorageClass)),
-				expectedSset:         withClaims(sampleSset, sampleClaim),
-				actualSset:           withClaims(sampleSset, sampleClaim),
-				validateStorageClass: true,
+				expectedSset: withClaims(sampleSset, sampleClaim),
+				actualSset:   withClaims(sampleSset, sampleClaim),
 			},
 			want: false,
-		},
-		{
-			name: "requested storage decrease: error out",
-			args: args{
-				k8sClient:            k8s.WrappedFakeClient(withVolumeExpansion(sampleStorageClass)),
-				expectedSset:         withClaims(sampleSset, sampleClaim, withStorageReq(sampleClaim2, "0.5Gi")),
-				actualSset:           withClaims(sampleSset, sampleClaim, sampleClaim2),
-				validateStorageClass: true,
-			},
-			want:    false,
-			wantErr: true,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := needsRecreate(tt.args.k8sClient, tt.args.expectedSset, tt.args.actualSset, tt.args.validateStorageClass)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("needsRecreate() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
+			got := needsRecreate(tt.args.expectedSset, tt.args.actualSset)
 			if got != tt.want {
 				t.Errorf("needsRecreate() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_isStorageExpansion(t *testing.T) {
-	type args struct {
-		expectedSize *resource.Quantity
-		actualSize   *resource.Quantity
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    bool
-		wantErr bool
-	}{
-		{
-			name: "expected == actual: false",
-			args: args{
-				expectedSize: resource.NewQuantity(1, resource.DecimalSI),
-				actualSize:   resource.NewQuantity(1, resource.DecimalSI),
-			},
-			want: false,
-		},
-		{
-			name: "expected > actual: true",
-			args: args{
-				expectedSize: resource.NewQuantity(2, resource.DecimalSI),
-				actualSize:   resource.NewQuantity(1, resource.DecimalSI),
-			},
-			want: true,
-		},
-		{
-			name: "expected < actual: error out",
-			args: args{
-				expectedSize: resource.NewQuantity(1, resource.DecimalSI),
-				actualSize:   resource.NewQuantity(2, resource.DecimalSI),
-			},
-			want:    false,
-			wantErr: true,
-		},
-		{
-			name: "expected is nil",
-			args: args{
-				expectedSize: nil,
-				actualSize:   resource.NewQuantity(1, resource.DecimalSI),
-			},
-			want: false,
-		},
-		{
-			name: "actual is nil",
-			args: args{
-				expectedSize: resource.NewQuantity(1, resource.DecimalSI),
-				actualSize:   nil,
-			},
-			want: false,
-		},
-		{
-			name: "expected and actual are nil",
-			args: args{
-				expectedSize: nil,
-				actualSize:   nil,
-			},
-			want: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := isStorageExpansion(tt.args.expectedSize, tt.args.actualSize)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("isStorageExpansion() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("isStorageExpansion() got = %v, want %v", got, tt.want)
 			}
 		})
 	}
@@ -713,218 +583,6 @@ func Test_removeESPodOwner(t *testing.T) {
 			require.NoError(t, err)
 			for i := range tt.wantPods {
 				comparison.RequireEqual(t, &tt.wantPods[i], &retrievedPods.Items[i])
-			}
-		})
-	}
-}
-
-func Test_ensureClaimSupportsExpansion(t *testing.T) {
-	tests := []struct {
-		name                string
-		k8sClient           k8s.Client
-		claim               corev1.PersistentVolumeClaim
-		validateStoragClass bool
-		wantErr             bool
-	}{
-		{
-			name:                "specified storage class supports volume expansion",
-			k8sClient:           k8s.WrappedFakeClient(withVolumeExpansion(sampleStorageClass)),
-			claim:               sampleClaim,
-			validateStoragClass: true,
-			wantErr:             false,
-		},
-		{
-			name:                "specified storage class does not support volume expansion",
-			k8sClient:           k8s.WrappedFakeClient(&sampleStorageClass),
-			claim:               sampleClaim,
-			validateStoragClass: true,
-			wantErr:             true,
-		},
-		{
-			name:                "default storage class supports volume expansion",
-			k8sClient:           k8s.WrappedFakeClient(withVolumeExpansion(defaultStorageClass)),
-			claim:               corev1.PersistentVolumeClaim{},
-			validateStoragClass: true,
-			wantErr:             false,
-		},
-		{
-			name:                "default storage class does not support volume expansion",
-			k8sClient:           k8s.WrappedFakeClient(&defaultStorageClass),
-			claim:               corev1.PersistentVolumeClaim{},
-			validateStoragClass: true,
-			wantErr:             true,
-		},
-		{
-			name:                "storage class vlaidation disabled: no-op",
-			k8sClient:           k8s.WrappedFakeClient(&sampleStorageClass), // would otherwise be refused: no expansion
-			claim:               sampleClaim,
-			validateStoragClass: false,
-			wantErr:             false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if err := ensureClaimSupportsExpansion(tt.k8sClient, tt.claim, tt.validateStoragClass); (err != nil) != tt.wantErr {
-				t.Errorf("ensureClaimSupportsExpansion() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
-func Test_allowsVolumeExpansion(t *testing.T) {
-	tests := []struct {
-		name string
-		sc   storagev1.StorageClass
-		want bool
-	}{
-		{
-			name: "allow volume expansion: true",
-			sc:   storagev1.StorageClass{AllowVolumeExpansion: pointer.BoolPtr(true)},
-			want: true,
-		},
-		{
-			name: "allow volume expansion: false",
-			sc:   storagev1.StorageClass{AllowVolumeExpansion: pointer.BoolPtr(false)},
-			want: false,
-		},
-		{
-			name: "allow volume expansion: nil",
-			sc:   storagev1.StorageClass{AllowVolumeExpansion: nil},
-			want: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := allowsVolumeExpansion(tt.sc); got != tt.want {
-				t.Errorf("allowsVolumeExpansion() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_isDefaultStorageClass(t *testing.T) {
-	tests := []struct {
-		name string
-		sc   storagev1.StorageClass
-		want bool
-	}{
-		{
-			name: "annotated as default",
-			sc:   defaultStorageClass,
-			want: true,
-		},
-		{
-			name: "annotated as default (beta)",
-			sc:   defaultBetaStorageClass,
-			want: true,
-		},
-		{
-			name: "annotated as default (+ beta)",
-			sc: storagev1.StorageClass{ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{
-				"storageclass.kubernetes.io/is-default-class":      "true",
-				"storageclass.beta.kubernetes.io/is-default-class": "true",
-			}}},
-			want: true,
-		},
-		{
-			name: "no annotations",
-			sc:   storagev1.StorageClass{ObjectMeta: metav1.ObjectMeta{Annotations: nil}},
-			want: false,
-		},
-		{
-			name: "not annotated as default",
-			sc:   sampleStorageClass,
-			want: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := isDefaultStorageClass(tt.sc); got != tt.want {
-				t.Errorf("isDefaultStorageClass() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_getDefaultStorageClass(t *testing.T) {
-	tests := []struct {
-		name      string
-		k8sClient k8s.Client
-		want      storagev1.StorageClass
-		wantErr   bool
-	}{
-		{
-			name:      "return the default storage class",
-			k8sClient: k8s.WrappedFakeClient(&sampleStorageClass, &defaultStorageClass),
-			want:      defaultStorageClass,
-		},
-		{
-			name:      "default storage class not found",
-			k8sClient: k8s.WrappedFakeClient(&sampleStorageClass),
-			want:      storagev1.StorageClass{},
-			wantErr:   true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := getDefaultStorageClass(tt.k8sClient)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("getDefaultStorageClass() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("getDefaultStorageClass() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_getStorageClass(t *testing.T) {
-	tests := []struct {
-		name      string
-		k8sClient k8s.Client
-		claim     corev1.PersistentVolumeClaim
-		want      storagev1.StorageClass
-		wantErr   bool
-	}{
-		{
-			name:      "return the specified storage class",
-			k8sClient: k8s.WrappedFakeClient(&sampleStorageClass, &defaultStorageClass),
-			claim:     corev1.PersistentVolumeClaim{Spec: corev1.PersistentVolumeClaimSpec{StorageClassName: pointer.StringPtr(sampleStorageClass.Name)}},
-			want:      sampleStorageClass,
-			wantErr:   false,
-		},
-		{
-			name:      "error out if not found",
-			k8sClient: k8s.WrappedFakeClient(&defaultStorageClass),
-			claim:     corev1.PersistentVolumeClaim{Spec: corev1.PersistentVolumeClaimSpec{StorageClassName: pointer.StringPtr(sampleStorageClass.Name)}},
-			want:      storagev1.StorageClass{},
-			wantErr:   true,
-		},
-		{
-			name:      "fallback to the default storage class if unspecified",
-			k8sClient: k8s.WrappedFakeClient(&sampleStorageClass, &defaultStorageClass),
-			claim:     corev1.PersistentVolumeClaim{Spec: corev1.PersistentVolumeClaimSpec{}},
-			want:      defaultStorageClass,
-			wantErr:   false,
-		},
-		{
-			name:      "error out if unspecified and default storage class not found",
-			k8sClient: k8s.WrappedFakeClient(&sampleStorageClass),
-			claim:     corev1.PersistentVolumeClaim{Spec: corev1.PersistentVolumeClaimSpec{}},
-			want:      storagev1.StorageClass{},
-			wantErr:   true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := getStorageClass(tt.k8sClient, tt.claim)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("getStorageClass() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !comparison.Equal(&got, &tt.want) {
-				t.Errorf("getStorageClass() got = %v, want %v", got, tt.want)
 			}
 		})
 	}
