+++ pkg/controller/elasticsearch/validation/validations.go
@@ -2,22 +2,32 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package v1
+package validation
 
 import (
+	"errors"
 	"fmt"
 	"net"
 	"strings"
 
 	commonv1 "github.com/elastic/cloud-on-k8s/pkg/apis/common/v1"
+	esv1 "github.com/elastic/cloud-on-k8s/pkg/apis/elasticsearch/v1"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/common/version"
 	esversion "github.com/elastic/cloud-on-k8s/pkg/controller/elasticsearch/version"
+	"github.com/elastic/cloud-on-k8s/pkg/utils/k8s"
 	netutil "github.com/elastic/cloud-on-k8s/pkg/utils/net"
+	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
+	storagev1 "k8s.io/api/storage/v1"
 	apiequality "k8s.io/apimachinery/pkg/api/equality"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/validation/field"
+	logf "sigs.k8s.io/controller-runtime/pkg/log"
 )
+
+var log = logf.Log.WithName("es-validation")
 
 const (
 	cfgInvalidMsg            = "Configuration invalid"
@@ -30,13 +40,13 @@ const (
 	nodeRolesInOldVersionMsg = "node.roles setting is not available in this version of Elasticsearch"
 	parseStoredVersionErrMsg = "Cannot parse current Elasticsearch version. String format must be {major}.{minor}.{patch}[-{label}]"
 	parseVersionErrMsg       = "Cannot parse Elasticsearch version. String format must be {major}.{minor}.{patch}[-{label}]"
-	pvcImmutableMsg          = "Volume claim templates can only have their storage requests increased, if the storage class allows volume expansion. Any other change is forbidden."
+	pvcImmutableErrMsg       = "volume claim templates can only have their storage requests increased, if the storage class allows volume expansion. Any other change is forbidden"
 	unsupportedConfigErrMsg  = "Configuration setting is reserved for internal use. User-configured use is unsupported"
 	unsupportedUpgradeMsg    = "Unsupported version upgrade path. Check the Elasticsearch documentation for supported upgrade paths."
 	unsupportedVersionMsg    = "Unsupported version"
 )
 
-type validation func(*Elasticsearch) field.ErrorList
+type validation func(esv1.Elasticsearch) field.ErrorList
 
 // validations are the validation funcs that apply to creates or updates
 var validations = []validation{
@@ -46,17 +56,7 @@ var validations = []validation{
 	supportedVersion,
 	validSanIP,
 }
-
-type updateValidation func(*Elasticsearch, *Elasticsearch) field.ErrorList
-
-// updateValidations are the validation funcs that only apply to updates
-var updateValidations = []updateValidation{
-	noDowngrades,
-	validUpgradePath,
-	pvcModification,
-}
-
-func (es *Elasticsearch) check(validations []validation) field.ErrorList {
+func check(es esv1.Elasticsearch, validations []validation) field.ErrorList {
 	var errs field.ErrorList
 	for _, val := range validations {
 		if err := val(es); err != nil {
@@ -67,20 +67,20 @@ func (es *Elasticsearch) check(validations []validation) field.ErrorList {
 }
 
 // noUnknownFields checks whether the last applied config annotation contains json with unknown fields.
-func noUnknownFields(es *Elasticsearch) field.ErrorList {
-	return commonv1.NoUnknownFields(es, es.ObjectMeta)
+func noUnknownFields(es esv1.Elasticsearch) field.ErrorList {
+	return commonv1.NoUnknownFields(&es, es.ObjectMeta)
 }
 
 // validName checks whether the name is valid.
-func validName(es *Elasticsearch) field.ErrorList {
+func validName(es esv1.Elasticsearch) field.ErrorList {
 	var errs field.ErrorList
-	if err := validateNames(es); err != nil {
+	if err := esv1.ValidateNames(es); err != nil {
 		errs = append(errs, field.Invalid(field.NewPath("metadata").Child("name"), es.Name, fmt.Sprintf("%s: %s", invalidNamesErrMsg, err)))
 	}
 	return errs
 }
 
-func supportedVersion(es *Elasticsearch) field.ErrorList {
+func supportedVersion(es esv1.Elasticsearch) field.ErrorList {
 	ver, err := version.Parse(es.Spec.Version)
 	if err != nil {
 		return field.ErrorList{field.Invalid(field.NewPath("spec").Child("version"), es.Spec.Version, parseVersionErrMsg)}
@@ -97,7 +97,7 @@ func supportedVersion(es *Elasticsearch) field.ErrorList {
 // The rules are:
 // There must be at least one master node.
 // node.roles are only supported on Elasticsearch 7.9.0 and above
-func hasCorrectNodeRoles(es *Elasticsearch) field.ErrorList {
+func hasCorrectNodeRoles(es esv1.Elasticsearch) field.ErrorList {
 	v, err := version.Parse(es.Spec.Version)
 	if err != nil {
 		return field.ErrorList{field.Invalid(field.NewPath("spec").Child("version"), es.Spec.Version, parseVersionErrMsg)}
@@ -112,8 +112,8 @@ func hasCorrectNodeRoles(es *Elasticsearch) field.ErrorList {
 	}
 
 	for i, ns := range es.Spec.NodeSets {
-		cfg := &ElasticsearchSettings{}
-		if err := UnpackConfig(ns.Config, *v, cfg); err != nil {
+		cfg := esv1.ElasticsearchSettings{}
+		if err := esv1.UnpackConfig(ns.Config, *v, &cfg); err != nil {
 			errs = append(errs, field.Invalid(confField(i), ns.Config, cfgInvalidMsg))
 
 			continue
@@ -143,39 +143,39 @@ func hasCorrectNodeRoles(es *Elasticsearch) field.ErrorList {
 	return errs
 }
 
-func getNodeRoleAttrs(cfg *ElasticsearchSettings) []string {
+func getNodeRoleAttrs(cfg esv1.ElasticsearchSettings) []string {
 	var nodeRoleAttrs []string
 
 	if cfg.Node != nil {
 		if cfg.Node.Data != nil {
-			nodeRoleAttrs = append(nodeRoleAttrs, NodeData)
+			nodeRoleAttrs = append(nodeRoleAttrs, esv1.NodeData)
 		}
 
 		if cfg.Node.Ingest != nil {
-			nodeRoleAttrs = append(nodeRoleAttrs, NodeIngest)
+			nodeRoleAttrs = append(nodeRoleAttrs, esv1.NodeIngest)
 		}
 
 		if cfg.Node.Master != nil {
-			nodeRoleAttrs = append(nodeRoleAttrs, NodeMaster)
+			nodeRoleAttrs = append(nodeRoleAttrs, esv1.NodeMaster)
 		}
 
 		if cfg.Node.ML != nil {
-			nodeRoleAttrs = append(nodeRoleAttrs, NodeML)
+			nodeRoleAttrs = append(nodeRoleAttrs, esv1.NodeML)
 		}
 
 		if cfg.Node.Transform != nil {
-			nodeRoleAttrs = append(nodeRoleAttrs, NodeTransform)
+			nodeRoleAttrs = append(nodeRoleAttrs, esv1.NodeTransform)
 		}
 
 		if cfg.Node.VotingOnly != nil {
-			nodeRoleAttrs = append(nodeRoleAttrs, NodeVotingOnly)
+			nodeRoleAttrs = append(nodeRoleAttrs, esv1.NodeVotingOnly)
 		}
 	}
 
 	return nodeRoleAttrs
 }
 
-func validSanIP(es *Elasticsearch) field.ErrorList {
+func validSanIP(es esv1.Elasticsearch) field.ErrorList {
 	var errs field.ErrorList
 	selfSignedCerts := es.Spec.HTTP.TLS.SelfSignedCertificate
 	if selfSignedCerts != nil {
@@ -191,7 +191,7 @@ func validSanIP(es *Elasticsearch) field.ErrorList {
 	return errs
 }
 
-func checkNodeSetNameUniqueness(es *Elasticsearch) field.ErrorList {
+func checkNodeSetNameUniqueness(es esv1.Elasticsearch) field.ErrorList {
 	var errs field.ErrorList
 	nodeSets := es.Spec.NodeSets
 	names := make(map[string]struct{})
@@ -208,49 +208,8 @@ func checkNodeSetNameUniqueness(es *Elasticsearch) field.ErrorList {
 	return errs
 }
 
-// pvcModification ensures the only part of volume claim templates that can be changed is storage requests.
-// VolumeClaimTemplates are immutable in StatefulSets, but we still manage to deal with volume expansion.
-func pvcModification(current, proposed *Elasticsearch) field.ErrorList {
+func noDowngrades(current, proposed esv1.Elasticsearch) field.ErrorList {
 	var errs field.ErrorList
-	if current == nil || proposed == nil {
-		return errs
-	}
-	for i, proposedNodeSet := range proposed.Spec.NodeSets {
-		currNodeSet := getNode(proposedNodeSet.Name, current)
-		if currNodeSet == nil {
-			// this is a new sset, so there is nothing to check
-			continue
-		}
-
-		// Check that no modification was made to the volumeClaimTemplates, except on storage requests.
-		// Checking semantic equality here allows providing PVC storage size with different units (eg. 1Ti vs. 1024Gi).
-		// We could conceptually also disallow storage decrease here (unsupported), but it would make it hard for users
-		// to revert to the previous size if tried increasing but their storage class doesn't support expansion.
-		currentClaims := currNodeSet.VolumeClaimTemplates
-		proposedClaims := proposedNodeSet.VolumeClaimTemplates
-		if !apiequality.Semantic.DeepEqual(claimsWithoutStorageReq(currentClaims), claimsWithoutStorageReq(proposedClaims)) {
-			errs = append(errs, field.Invalid(field.NewPath("spec").Child("nodeSet").Index(i).Child("volumeClaimTemplates"), proposedClaims, pvcImmutableMsg))
-		}
-	}
-	return errs
-}
-
-// claimsWithoutStorageReq returns a copy of the given claims, with all storage requests set to the empty quantity.
-func claimsWithoutStorageReq(claims []corev1.PersistentVolumeClaim) []corev1.PersistentVolumeClaim {
-	result := make([]corev1.PersistentVolumeClaim, 0, len(claims))
-	for _, claim := range claims {
-		patchedClaim := *claim.DeepCopy()
-		patchedClaim.Spec.Resources.Requests[corev1.ResourceStorage] = resource.Quantity{}
-		result = append(result, patchedClaim)
-	}
-	return result
-}
-
-func noDowngrades(current, proposed *Elasticsearch) field.ErrorList {
-	var errs field.ErrorList
-	if current == nil || proposed == nil {
-		return errs
-	}
 	currentVer, err := version.Parse(current.Spec.Version)
 	if err != nil {
 		// this should not happen, since this is the already persisted version
@@ -269,11 +228,8 @@ func noDowngrades(current, proposed *Elasticsearch) field.ErrorList {
 	return errs
 }
 
-func validUpgradePath(current, proposed *Elasticsearch) field.ErrorList {
+func validUpgradePath(current, proposed esv1.Elasticsearch) field.ErrorList {
 	var errs field.ErrorList
-	if current == nil || proposed == nil {
-		return errs
-	}
 	currentVer, err := version.Parse(current.Spec.Version)
 	if err != nil {
 		// this should not happen, since this is the already persisted version
@@ -299,12 +255,113 @@ func validUpgradePath(current, proposed *Elasticsearch) field.ErrorList {
 	}
 	return errs
 }
+func ValidateClaimsUpdate(
+func ValidateClaimsStorageUpdate(
+	k8sClient k8s.Client,
+	initial []corev1.PersistentVolumeClaim,
+	updated []corev1.PersistentVolumeClaim,
+	validateStorageClass bool,
+) error {
+	for _, updatedClaim := range updated {
+		initialClaim := claimMatchingName(initial, updatedClaim.Name)
+		if initialClaim == nil {
+			// existing claim does not exist in updated
+			return errors.New(pvcImmutableErrMsg)
+		}
 
-func getNode(name string, es *Elasticsearch) *NodeSet {
-	for i := range es.Spec.NodeSets {
-		if es.Spec.NodeSets[i].Name == name {
-			return &es.Spec.NodeSets[i]
+		cmp := k8s.CompareStorageRequests(initialClaim.Spec.Resources, updatedClaim.Spec.Resources)
+		switch {
+		case cmp.Increase:
+			// storage increase requested: ensure the storage class allows volume expansion
+			if err := EnsureClaimSupportsExpansion(k8sClient, updatedClaim, validateStorageClass); err != nil {
+				return err
+			}
+		case cmp.Decrease:
+			// storage decrease is not supported
+			return fmt.Errorf("decreasing storage size is not supported: an attempt was made to decrease storage size for claim %s", updatedClaim.Name)
 		}
 	}
 	return nil
+}
+
+func claimMatchingName(claims []corev1.PersistentVolumeClaim, name string) *corev1.PersistentVolumeClaim {
+	for i, claim := range claims {
+		if claim.Name == name {
+			return &claims[i]
+		}
+	}
+	return nil
+}
+
+// claimsWithoutStorageReq returns a copy of the given claims, with all storage requests set to the empty quantity.
+func claimsWithoutStorageReq(claims []corev1.PersistentVolumeClaim) []corev1.PersistentVolumeClaim {
+	result := make([]corev1.PersistentVolumeClaim, 0, len(claims))
+	for _, claim := range claims {
+		patchedClaim := *claim.DeepCopy()
+		patchedClaim.Spec.Resources.Requests[corev1.ResourceStorage] = resource.Quantity{}
+		result = append(result, patchedClaim)
+	}
+	return result
+}
+
+// EnsureClaimSupportsExpansion inspects whether the storage class referenced by the claim
+// allows volume expansion, and returns an error if it doesn't.
+func EnsureClaimSupportsExpansion(k8sClient k8s.Client, claim corev1.PersistentVolumeClaim, validateStorageClass bool) error {
+	if !validateStorageClass {
+		log.V(1).Info("Skipping storage class validation")
+		return nil
+	}
+	sc, err := getStorageClass(k8sClient, claim)
+	if err != nil {
+		return err
+	}
+	if !allowsVolumeExpansion(sc) {
+		return fmt.Errorf("claim %s (storage class %s) does not support volume expansion", claim.Name, sc.Name)
+	}
+	return nil
+}
+
+// getStorageClass returns the storage class specified by the given claim,
+// or the default storage class if the claim does not specify any.
+func getStorageClass(k8sClient k8s.Client, claim corev1.PersistentVolumeClaim) (storagev1.StorageClass, error) {
+	if claim.Spec.StorageClassName == nil || *claim.Spec.StorageClassName == "" {
+		return getDefaultStorageClass(k8sClient)
+	}
+	var sc storagev1.StorageClass
+	if err := k8sClient.Get(types.NamespacedName{Name: *claim.Spec.StorageClassName}, &sc); err != nil {
+		return storagev1.StorageClass{}, fmt.Errorf("cannot retrieve storage class: %w", err)
+	}
+	return sc, nil
+}
+
+// getDefaultStorageClass returns the default storage class in the current k8s cluster,
+// or an error if there is none.
+func getDefaultStorageClass(k8sClient k8s.Client) (storagev1.StorageClass, error) {
+	var scs storagev1.StorageClassList
+	if err := k8sClient.List(&scs); err != nil {
+		return storagev1.StorageClass{}, err
+	}
+	for _, sc := range scs.Items {
+		if isDefaultStorageClass(sc) {
+			return sc, nil
+		}
+	}
+	return storagev1.StorageClass{}, errors.New("no default storage class found")
+}
+
+// isDefaultStorageClass inspects the given storage class and returns true if it is annotated as the default one.
+func isDefaultStorageClass(sc storagev1.StorageClass) bool {
+	if len(sc.Annotations) == 0 {
+		return false
+	}
+	if sc.Annotations["storageclass.kubernetes.io/is-default-class"] == "true" ||
+		sc.Annotations["storageclass.beta.kubernetes.io/is-default-class"] == "true" {
+		return true
+	}
+	return false
+}
+
+// allowsVolumeExpansion returns true if the given storage class allows volume expansion.
+func allowsVolumeExpansion(sc storagev1.StorageClass) bool {
+	return sc.AllowVolumeExpansion != nil && *sc.AllowVolumeExpansion
 }
